requirements.txt
aiohttp==3.6.2
aiounittest==1.3.1
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
coverage==5.1
discord.py==1.3.3
docopt==0.6.2
idna==2.9
multidict==4.7.5
PyYAML==5.3.1
testfixtures==6.14.1
websockets==8.1
yarl==1.4.2
